+++ liblua/lbaselib.c
@@ -133,6 +133,15 @@ static int luaB_getfenv (lua_State *L) {
 static int luaB_setfenv (lua_State *L) {
     luaL_checktype(L, 2, LUA_TTABLE);
     getfunc(L, 0);
+
+    lua_getfenv(L, -1);
+
+    if (luaL_getmetafield(L, -1, "__environment")) {
+        luaL_error(L, "cannot change a protected environment");
+    } else {
+        lua_pop(L, 1);
+    }
+
     lua_taintobject(L, -1);
     lua_pushvalue(L, 2);
     if (lua_isnumber(L, 1) && lua_tonumber(L, 1) == 0) {